Part 1 Complete (ready for Parts 2-4)
**What was updated**
• Fixed styling issues.
• *NOTE* you may have to hold Shift+Refresh for updates in main.css to
appear.
• Fixed all functional aspects and added a bit more error checking.
• Added a products.bak file, just in case. Rename to products.txt if you
need a fresh copy of the products.txt file.
**Known Bug**
• When updating or deleting a product, the prodList won't show the
updated status unless you refresh the page, or leave and return to the
page.
•• I believe this is caused by a race condition. The ProdIO add/delete
does not complete updates to the text file *before* the prodList
redirects and reads the file.
•• I will look into this further after Part 3 is complete.  It won't
block progress, and once Part3 is implemented, this bug will probably go
away.